rust: satisfy structural limits in connector delivery/tests
